pyproject.toml
[tool.poetry]
name = "cloudguard-modules"
version = "1.0.3"
description = "Cloudguard provider modules written in python to be used in cloudguard CLI"
authors = ["deepraj1729 <deeprajbaidyaofficial@gmail.com>", "Saptaktdk <saptakdev123@gmail.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/limebrew-org/cloudguard-modules"
repository = "https://github.com/limebrew-org/cloudguard-modules"
keywords = ["python","aws","azure","gcp","cspm"]
packages = [{include = "cloudguard_modules"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.95.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
google-auth = "^2.17.1"
google-cloud-compute = "1.13.0"
google-cloud-dns = "0.34.1"
google-cloud-storage = "2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"